# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interview challenge for building a Payment Gateway - Go version",
        "title": "Payment Gateway Challenge Go",
        "contact": {}
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
        "/api/payments": {
            "post": {
                "description": "Requests a payment be sent to a simulated bank API. Persists the outcome of the payment and  returns the processed payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Process Payment",
                "operationId": "request-payment",
                "parameters": [
                    {
                        "description": "Payment Process Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProcessPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "description": "Gets a Payment by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get a Payment",
                "operationId": "get-payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the payment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Payment"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ping"
                ],
                "summary": "Ping the server",
                "operationId": "get-ping",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "card_number_last_four": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "expiry_month": {
                    "type": "integer"
                },
                "expiry_year": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "payment_status": {
                    "$ref": "#/definitions/status.State"
                }
            }
        },
        "models.ProcessPaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "card_number": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "cvv": {
                    "type": "string"
                },
                "expiry_month": {
                    "type": "integer"
                },
                "expiry_year": {
                    "type": "integer"
                }
            }
        },
        "status.State": {
            "type": "string",
            "enum": [
                "Authorized",
                "Declined"
            ],
            "x-enum-varnames": [
                "StateAuthorized",
                "StateDeclined"
            ]
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}